Add support for using as a custom merge driver
This PR adds enough support to use merge-fmt as a git custom merge
driver. This is basically just support for passing the filename of the
final merge result to the formatter so that the correct config can be
used (and to avoid problems due to the absence of file extensions on
the temporary files).

This requires https://github.com/ocaml-ppx/ocamlformat/pull/740 so
that ocamlformat will accept `--name` with `--inplace`.

See
https://git-scm.com/docs/gitattributes#_defining_a_custom_merge_driver
for the docs on how to configure this, but basically adding
```
[merge "mergefmt"]
	name = merge-fmt driver
	driver = merge-fmt mergetool --base=%O --current=%A --other=%B -o %A --name=%P
[merge]
	tool = mergefmt
	default = mergefmt
```
to `.git/config` causes regular merge operations done by git, such as
when rebasing, to use `merge-fmt`.